migration fix

+++ db/migrate/20230316221905_remove_column_from_users.rb
@@ -1,5 +1,5 @@
 class RemoveColumnFromUsers < ActiveRecord::Migration[7.0]
   def change
-    remove_column :users, :document, :string
+    remove_column :users, :document, :string if column_exists?(:users, :document)
   end
 end
